trademark_ocr_notice/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==1.5.0
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
cheroot==8.3.0
cnocr==1.2.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
future==0.18.2
gast==0.3.3
gluoncv==0.6.0
graphviz==0.8.4
grpcio==1.29.0
h5py==2.10.0
html5lib==0.9999999
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jaraco.functools==3.0.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.0.1
kiwisolver==1.2.0
lmdb==0.98
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.3.0
mxnet==1.6.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.19.0
opencv-contrib-python==4.0.0.21
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==1.7.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pyee==7.0.2
PyExecJS==1.5.1
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
redis==3.5.3
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
Theano==1.0.4
threadpoolctl==2.1.0
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
web.py==0.40.dev0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
